// src/http-busca-search/models/getApiV1CategoriesCategorySubcategoriesParams.ts
import type { GetApiV1CategoriesCategorySubcategoriesOrder } from './getApiV1CategoriesCategorySubcategoriesOrder'
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Mecanismo de Busca API
 * API para busca textual e vetorial usando Typesense e embeddings gerados via Google Gemini
 * OpenAPI spec version: 1.0
 */
import type { GetApiV1CategoriesCategorySubcategoriesSortBy } from './getApiV1CategoriesCategorySubcategoriesSortBy'

export type GetApiV1CategoriesCategorySubcategoriesParams = {
  /**
   * Critério de ordenação
   */
  sort_by?: GetApiV1CategoriesCategorySubcategoriesSortBy
  /**
   * Direção da ordenação
   */
  order?: GetApiV1CategoriesCategorySubcategoriesOrder
  /**
   * Incluir subcategorias sem serviços publicados
   */
  include_empty?: boolean
  /**
   * Incluir serviços inativos/rascunhos (status != 1) nas contagens
   */
  include_inactive?: boolean
}